Adds Laravel Dusk for browser testing
This commit introduces the Laravel Dusk package for handling end-to-end browser testing. Several related configuration files and tests were added, including a new .env file and DuskTestCase. Improvements were also made to the recipe fetching process by updating the fetchAll and fetch methods in the 'useRecipeStore.ts' file and adding meta details in the 'usePaginationStore.ts' file. Lastly, the visual representation of the total number of fetched recipes was enhanced in 'index.vue'.