Expands seeders & adds more Dusk tests
Added provisions in the IngredientFactory for localizing the name of a common ingredient. Recipe tests were expanded to cover pagination, search parameters, and show page functionality in detail. Also, enhancements were made in the seed data for testing more diverse cases in the TestRecipesSeeder and a MySQL connection specifically for tests was added.